refactor run.py.
